Update getting-started.rst
When dealing with a multi currency site, FE, it can drive you nuts to understand Money initialization if you don't realice you've to create the object represented in the smallest unit. The examples in the documentation use EUR and USD as example currencies, and both hace 2 minor units. I think it'd be very beneficial to state that other currencies may have other minor units, or none!

It took me 5 hours to notice it :)