Upgraded sources to PHP 8 type declarations, reworked internal number logic to always work on `string` values
Here we introduce a lot of BC breaks in type signatures, specifically calculators. Values that used to have `float`
as acceptable state are now generally changed to `numeric-string`, although not everywhere (yet).
This is not yet the final form of the codebase, but we need an atomic state where:

 * we are at a newer/better CS level (specifically psalm type declarations and native type declarations)
 * we have a passing PHPUnit/PHPSpec test suite

By changing this, we gain a lot of internal consistency that we can leverage to easier refactor components
that need to do a lot of intensive performance-oriented math operations.

Note that we are also introducing a lot of type contradiction (stricter types, but we still check for input
value shape): that is because we did not yet upgrade `vimeo/psalm`, and therefore we haven't yet tackled
any of these, but rather based our changes on the existing test suite instead.