Brought type inference to 100% of the codebase (tests still missing)
Here we change the philosophy of the library radically: instead of doing runtime
validation for inputs, we leave those up to the user (yes, really! endorsing a type-checker
will need to be part of this patch), and instead we expect correct values to be passed
in as type-checked upfront.

This way, we get rid of a number of runtime validation tests that are completely avoidable.